ADM-369 move general style into ReportStep
